% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buttongroup.R
\name{buttonGroup}
\alias{buttonGroup}
\title{Creates a button group component}
\usage{
buttonGroup(inputId, lstItems, selectedItems = list(),
  activeColor = "green", inactiveColor = "light gray",
  impliedColor = "orange", label = "")
}
\arguments{
\item{inputId}{id for the button group component. The selected groups will be returned to input$inputId}

\item{lstItems}{list with all items to display in the buttun group}

\item{selectedItems}{items which will be initialy selected in the button group}

\item{activeColor}{color for selected items}

\item{inactiveColor}{color for inactive items}

\item{impliedColor}{color for the special status 'implied' This status is uesfull in the context of the filter app.
       In the filter app there will be multiple button groups and certain selections can imply the absence of items on other groups.}
}
\value{
A button group input control that can be added to a UI definition.
}
\description{
Creates a button group component
}
\examples{
buttonGroup('btnGroup',list('A','B','C'),selectedItems=list('A','B'))
}
\seealso{
Other buttongroup: \code{\link{resetButtonGroup}};
  \code{\link{updateButtonGroup}}
}

