% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buttongroup.R
\name{updateButtonGroup}
\alias{updateButtonGroup}
\title{Send an update message to a URL input on the client.
This update message can change the value and/or label.}
\usage{
updateButtonGroup(session, inputId, selectedItems = NULL,
  impliedItems = NULL)
}
\arguments{
\item{session}{shiny session object}

\item{inputId}{Id of the buttongroup to update}

\item{selectedItems}{items to set the selected status for}

\item{impiedItems}{items to set the implied status for}
}
\description{
Send an update message to a URL input on the client.
This update message can change the value and/or label.
}
\examples{
updateButtonGroup(session,'btnGroup',selectedItems=list('C'),impliedItems=list('A','B'))
}
\seealso{
Other buttongroup: \code{\link{buttonGroup}};
  \code{\link{resetButtonGroup}}
}

